backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
websockets
openai
aiofiles
python-socketio
email-validator
pandas
numpy

